Add support for ssh+chroot.
Add a `chroot_dir` argument to both play and task in order to run them
in a remote chroot environment.

This allows for example to configure debootstrap on a remote host.

Example:

`````
  - hosts: "all"
    chroot_dir: /srv/chroot

    tasks:
    - name: copy foo
      copy: src=/tmp/foo dest=/root/foo
`````

Will copy `/tmp/foo` into `/root/foo` on remote hosts in the
`/srv/chroot` environment. The task is not executed on the remote host
itself, but in the `/srv/chroot` environment on the remote host.

This permits to use native ansible module in a remote chroot environment.

The `chroot_dir` precedence is:

 - at the play level
 - at the task level

So one task can be run in an other chroot if needed.

Please note that `chroot_dir` is only compatible with the `ssh`
connection. Meanwhile a `sudo` can be done.